# shared/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "EnrolRequest": {
        "additionalProperties": false,
        "properties": {
          "device_label": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          },
          "join_code": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Join Code",
            "type": "string"
          },
          "mode": {
            "default": "FIELD",
            "pattern": "^(FIELD|GATEWAY)$",
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "join_code"
        ],
        "title": "EnrolRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JoinCodeRequest": {
        "additionalProperties": false,
        "properties": {
          "incident_id": {
            "minimum": 0.0,
            "title": "Incident Id",
            "type": "integer"
          },
          "max_uses": {
            "default": 5,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Uses",
            "type": "integer"
          }
        },
        "required": [
          "incident_id"
        ],
        "title": "JoinCodeRequest",
        "type": "object"
      },
...
      "SessionRequest": {
        "additionalProperties": false,
        "properties": {
          "incident_id": {
            "minimum": 0.0,
            "title": "Incident Id",
            "type": "integer"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "password"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Multi-agency search-and-rescue coordination. REST is authoritative for all durable writes and reads.",
    "title": "RescueNet AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/session": {
      "post": {
        "description": "Authenticated by the incident password (`09` endpoint 2).",
        "operationId": "open_session_v1_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Open Session V1 Auth Session Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commander dashboard login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/enrol": {
      "post": {
        "description": "Authenticated by the join code itself \u2014 no bearer token (`09` endpoint 1).\n\nDeliberately **non-idempotent**: two enrolments with one code yield two\ndistinct `device_id` values (AP-02, TR-33).",
        "operationId": "enrol_v1_enrol_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enrol V1 Enrol Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device joins an incident with a join code",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/teams/{team_id}/join-code": {
      "post": {
        "operationId": "mint_join_code_v1_teams__team_id__join_code_post",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Mint Join Code V1 Teams  Team Id  Join Code Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mint a join code for a team",
        "tags": [
          "auth"
        ]
      }
    }
  }
}